Provided some dependencies

+++ app/models/db/User.scala
@@ -3,7 +3,7 @@ package models.db
 /**
   * Created by pozpl on 03.04.17.
   */
-case class User(id: Option[Int], uid: String, login: String, mobile:Option[Long], email: Option[String])
+case class User(id: Option[Long], uid: String, login: String, mobile:Option[Long], email: Option[String])
